Reports: forecast query sortAsc = true
